+++ enet-csharp/ENet/define/system.cs
@@ -1,7 +1,7 @@
 ﻿using System.Diagnostics;
 using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
-using size_t = nint;
+using size_t = nuint;
 using enet_uint8 = byte;
 
 // ReSharper disable ALL
@@ -15,7 +15,7 @@ namespace enet
 #if NET6_0_OR_GREATER
             return NativeMemory.Alloc((nuint)size);
 #else
-            return (void*)Marshal.AllocHGlobal(size);
+            return (void*)Marshal.AllocHGlobal((nint)size);
 #endif
         }
